+++ ArcheAgeLogin/ArcheAge/Network/GameServerPacket.cs
@@ -34,7 +34,7 @@ namespace ArcheAgeLogin.ArcheAge.Network
                     /*                 accountId Lv Ms       Name               Sesion   LastEnteredTime    LastIp
                      * Recv: 6C00 0100 10000000  00 00 0000 333632383338333030 00396166 373166343463656336 31 62 31 36 35 37 32 38 64 38 63 36 65 34 32 34 30 62 33 39 38 33 30 63 65 61 61 66 30 37 66 31 35 35 36 64 62 32 65 64 33 30 32 61 38 64 39 39 37 35 00 01 01 31 32 37 2E 30 2E 30 2E 31 00 21 E5 F6 81 65 01 00 00 00 4F 8E FC D4
                      */
-                    ns.Write((int)account.AccountId);
+                    ns.Write((uint)account.AccountId);
                     ns.WriteDynamicASCII(account.Name);
                     //ns.WriteDynamicASCII(account.Password);
                     ns.WriteDynamicASCII(account.Token);
@@ -50,7 +50,7 @@ namespace ArcheAgeLogin.ArcheAge.Network
                     /*                 accountId        Lv Ms Name           Sesion   LastEnteredTime  LastIp
                      * Send: 2900 0100 1AC7000000000000 01 01 61617465737400 2810B47A 565074D264010000 3132372E302E302E3100
                      */
-                    ns.Write((long)account.AccountId);
+                    ns.Write((uint)account.AccountId);
                     ns.WriteDynamicASCII(account.Name);
                     //ns.WriteDynamicASCII(account.Password);
                     ns.WriteDynamicASCII(account.Token);
